@@ -1,6 +1,6 @@
 {
   "name": "@hcikit/react",
-  "version": "1.0.0-alpha.11",
+  "version": "1.0.0-alpha.13",
   "description": "An experiment runner with a set of pre-built experiment components for running hci experiments",
   "author": "blainelewis1",
   "license": "MIT",
@@ -42,17 +42,17 @@
     "@mui/icons-material": "^5.5.0",
     "@mui/material": "^5.5.0",
     "@mui/styles": "^5.5.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "dependencies": {
     "@hcikit/workflow": "workspace:*",
     "core-js": "^3.26.1",
     "lodash-es": "^4.17.21",
-    "markdown-to-jsx": "^7.1.7",
     "prop-types": "^15.8.1",
     "react-error-boundary": "^3.1.4",
     "react-is": "^18.2.0",
+    "react-markdown": "^8.0.3",
     "serialize-error": "^11.0.0",
     "styled-components": "^5.3.6"
   },
@@ -63,6 +63,7 @@
     "@types/jest": "^29.2.3",
     "@types/react": "^18.0.25",
     "@types/styled-components": "^5.1.26",
+    "@types/testing-library__jest-dom": "^5.14.5",
     "jest-diff": "^29.3.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
