@@ -1,6 +1,10 @@
 import React from "react";
 import { Paper } from "@mui/material";
-import styled from "styled-components";
+import _styled from "styled-components";
+
+const styled: typeof _styled =
+  ((_styled as unknown as { default: typeof _styled })
+    .default as typeof _styled) || _styled;
 
 const Overlay = styled.div`
   position: absolute;
